AP-5336: Refactor state benefit analyser
Doing to much and this can be extract to a loading/setup
class easily enough.
